requirements.txt
torch>=2.0.0
numpy>=1.23.5
hydra-core>=1.3.2
omegaconf>=2.3.0
plotly>=5.13.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=1.5.3
resampy>=0.4.3
librosa>=0.11.0
h5py>=3.15.1
